[chore][pkg/stanza] Cleanup json parser operator files
