+++ pkg/stanza/operator/parser/json/parser_test.go
